llm_internship_starter/requirements.txt
streamlit
pandas
lightweight-charts-python
google-generativeai
